이미지한블럭씩 사라짐-위쪽만 공 튀겨짐,충돌범위 안지워짐

+++ FrameWork/Blocks.h
@@ -8,7 +8,7 @@ class Blocks
 		int point;				//충돌했을때 먹는 점수
 		bool life;
 		bool death;
-		double dx, dy, dz, drx, dry, drz;	// 데미지존
+		float dx, dy, dz, drx, dry, drz;	// 데미지존
 	};
 
 	Sprite Bkimg[5];	//이미지
